rework abstract
IMO a good abstract must glue together the paper title and
its conclusions. The current abstract is
1) Too long
2) Misleading: puts the reader's attention on the tsi148,
   whereas this is all about the bigger picture (of course we
   make our point with the tsi148 driver)
3) One has a wtf moment right after the abstract ends, with
   the Section on the TSI148, which IMO appears out of nowhere.
   The abstract does not prepare the reader for it.

I've humbly tried to fix the above points, although the Real Fix
would probably be an introductory section, with the usual last
paragraph "the paper is structured as follows". But forget that
crap, I like the fact that we go straight to the point. So I've
compressed the hypothetical introduction into the abstract. Hmmm?

Signed-off-by: Emilio G. Cota <cota@braap.org>